Add a WeatherDay struct and declare a metatype
This adds a WeatherDay structure to the code which is intended to be
used by synchronization apps for the weather data.  The internal
structure is three parallel arrays of 5 items each containing the icon
descriptor, the low temperature and high temperature for each day.  This
leaves that interface in place but adds a somewhat more usable array of
such triplets.