Patch screenshot service to avoid crashing
Sometimes the first packet received back from the watch is not the
length or is not four bytes long.  This fixes both issues by explicitly
setting variables in during construction of the service object.  It also
fixes a latent bug that would cause incorrect size setting if any of
the size bytes were greater than 0x7f and therefore interpreted as a
signed integer and incorrectly extending the sign bit.

Signed-off-by: Ed Beroset <beroset@ieee.org>