Use WriteWithResponse for longer messages
This fixes https://github.com/AsteroidOS/libasteroid/issues/21 by only
using WriteWitoutResponse for two messages: the time set message and
the screenshot request message.  The reason is that the Qt documentation
states that if WriteWithoutResponse is used, the payload can be no more
than 20 bytes but this is often not the case for most of these messages.

Signed-off-by: Ed Beroset <beroset@ieee.org>